requirements.txt
apscheduler==3.11.0
pandas==2.3.0
psycopg2-binary==2.9.10
pytz==2024.1
requests==2.32.4
schedule==1.2.2